refactor ok
