+++ transcendence/base/templates/base.html
@@ -103,5 +103,8 @@
   <script src="{% static 'js/nav.js' %}" type="module" defer></script>
   <script src="{% static 'js/api.js' %}" type="module" defer></script>
   <script src="{% static 'js/auth.js' %}" type="module" defer></script>
-  <script src="{% static 'js/pongvs4.js' %}" type="module" defer></script>
+  <script src="{% static 'js/pong/pong_light.js' %}" type="module" defer></script>
+  <script src="{% static 'js/pong/pong_var.js' %}" type="module" defer></script>
+  <script src="{% static 'js/pong/pong_obj.js' %}" type="module" defer></script>
+  <script src="{% static 'js/pong/pongvs4.js' %}" type="module" defer></script>
 </html>
