+++ transcendence/base/templates/play.html
@@ -1,5 +1,10 @@
 <div id="root">
 	<div id="pong">
+		<script src="../static/js/pong/pong_var.js" type="module" defer></script>
+		<script src="../static/js/pong/pong_obj.js" type="module" defer></script>
+		<script src="../static/js/pong/pong_light.js" type="module" defer></script>
+		<script src="../static/js/pong/pongvs4.js" type="module" defer></script>
+		<canvas id="game"></canvas>
 		<div id="playerInfo1">
 		<div id="p1Score">0</div>
 		<div id="player1avatar">
